+++ tests/qprog/qprog_contracts.py
@@ -21,7 +21,7 @@ OPTIMIZERS_TO_TEST = {
         lambda: ScipyOptimizer(method=ScipyMethod.L_BFGS_B),
         lambda: ScipyOptimizer(method=ScipyMethod.COBYLA),
         lambda: ScipyOptimizer(method=ScipyMethod.NELDER_MEAD),
-        lambda: PymooOptimizer(method=PymooMethod.CMAES, population_size=5),
+        lambda: PymooOptimizer(method=PymooMethod.CMAES, population_size=10),
         lambda: PymooOptimizer(method=PymooMethod.DE, population_size=5),
     ],
     "ids": ["MonteCarlo", "L_BFGS_B", "COBYLA", "NELDER_MEAD", "CMAES", "DE"],